organize files into directories
