+++ bin/cli.js
@@ -1,12 +1,12 @@
 #!/usr/bin/env node
 const [,, ...args] = process.argv;
-const createLambda = require('./createLambda.js');
-const init = require('./init.js');
-const createDirectory = require('./createDirectory.js');
-const createJSONFile = require('./createJSONFile.js');
-const configTemplate = require('./configTemplate.js');
-const deployLambda = require('./deployLambda.js');
-const asyncQuestion = require('./asyncQuestion.js');
+const createLambda = require('../src/commands/createLambda.js');
+const init = require('../src/commands/init.js');
+const createDirectory = require('../src/util/createDirectory.js');
+const createJSONFile = require('../src/util/createJSONFile.js');
+const configTemplate = require('../src/util/configTemplate.js');
+const deployLambda = require('../src/commands/deployLambda.js');
+const asyncQuestion = require('../src/util/asyncQuestion.js');
 
 const defaultRole = 'defaultBamRole';
 const configJSON = configTemplate(defaultRole);
@@ -24,7 +24,7 @@ const configJSON = configTemplate(defaultRole);
     // prompt user for decription
     const description = await asyncQuestion('Please give a brief description of your lambda: ', '');
     deployLambda(args[2], description, '.');
-  } else {  
+  } else {
 
   }
 })();
