+++ src/util/asyncQuestion.js
@@ -2,10 +2,10 @@ const readline = require('readline');
 
 const rl = readline.createInterface({
   input: process.stdin,
-  output: process.stdout
+  output: process.stdout,
 });
 
-module.exports = async function asyncQuestion (prompt, defaultValue) {
+module.exports = async function asyncQuestion(prompt, defaultValue) {
   const question = new Promise((resolve, reject) => {
     rl.question(prompt, resolve);
     rl.write(defaultValue);
@@ -14,4 +14,3 @@ module.exports = async function asyncQuestion (prompt, defaultValue) {
   rl.close();
   return answer;
 };
-
