+++ src/util/configTemplate.js
@@ -1,4 +1,3 @@
-const fs = require('fs');
 const getRegion = require('./getRegion.js');
 
 module.exports = function configTemplate(roleName) {
