+++ src/util/doesResourceExist.js
