+++ test/init.test.js
@@ -7,11 +7,11 @@ const { promisify } = require('util');
 const AWS = require('aws-sdk');
 const fs = require('fs');
 
-const createDirectory = require('../createDirectory.js');
-const createJSONFile = require('../createJSONFile.js');
-const { doesRoleExist, doesPolicyExist } = require('../doesResourceExist.js');
-const createRole = require('../createRole.js');
-const configTemplate = require('../configTemplate.js');
+const createDirectory = require('../src/util/createDirectory.js');
+const createJSONFile = require('../src/util/createJSONFile.js');
+const { doesRoleExist, doesPolicyExist } = require('../src/util/doesResourceExist.js');
+const createRole = require('../src/util/createRole.js');
+const configTemplate = require('../src/util/configTemplate.js');
 
 const iam = new AWS.IAM();
 const roleName = 'testDefaultBamRole';
